// client/stocklinx app/src/redux/category/constant.ts
export enum categoryConst {
  FETCH_CATEGORIES_REQUEST = "FETCH_CATEGORIES_REQUEST",
  FETCH_CATEGORIES_SUCCESS = "FETCH_CATEGORIES_SUCCESS",
  FETCH_CATEGORIES_FAILURE = "FETCH_CATEGORIES_FAILURE",
  FETCH_CATEGORY_REQUEST = "FETCH_CATEGORY_REQUEST",
  FETCH_CATEGORY_SUCCESS = "FETCH_CATEGORY_SUCCESS",
  FETCH_CATEGORY_FAILURE = "FETCH_CATEGORY_FAILURE",
  CREATE_CATEGORY_REQUEST = "CREATE_CATEGORY_REQUEST",
  CREATE_CATEGORY_SUCCESS = "CREATE_CATEGORY_SUCCESS",
  CREATE_CATEGORY_FAILURE = "CREATE_CATEGORY_FAILURE",
  UPDATE_CATEGORY_REQUEST = "UPDATE_CATEGORY_REQUEST",
  UPDATE_CATEGORY_SUCCESS = "UPDATE_CATEGORY_SUCCESS",
  UPDATE_CATEGORY_FAILURE = "UPDATE_CATEGORY_FAILURE",
  REMOVE_CATEGORY_REQUEST = "REMOVE_CATEGORY_REQUEST",
  REMOVE_CATEGORY_SUCCESS = "REMOVE_CATEGORY_SUCCESS",
  REMOVE_CATEGORY_FAILURE = "REMOVE_CATEGORY_FAILURE",
  SET_CATEGORY = "SET_CATEGORY",
  CLEAR_CATEGORY = "CLEAR_CATEGORY",
  SET_CATEGORIES = "SET_CATEGORIES",
  CLEAR_CATEGORIES = "CLEAR_CATEGORIES",
}